Moved soc_u.* to soc/soc_u.*
