+++ src/core/hle/service/soc/soc_u.h
